% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adam_grpreg}
\alias{adam_grpreg}
\title{The ADAM Algorithm for GLM with gLASSO, gSCAD, gMCP penalties (experimental)}
\usage{
adam_grpreg(X, y, beta, lambda, a, group, family, penalty, stepsize, tol,
  maxit, b1 = 0.7, b2 = 0.9, e = 1e-08)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{The response vector.}

\item{beta}{The model coefficients.}

\item{lambda}{The penalization parameter for gMIC, e.g., 2 for AIC and long(n) for BIC.}

\item{a}{The approximation parameter for MCP or SCAD.}

\item{group}{The group structure of the model. For example, assume that X has 4 columns and group=c(1,1,2,2).
It means the first 2 features form a group of variables and the last 2 features form another group of variables.}

\item{family}{The type of glm model, should be one of "gaussian", "binomial" or "poisson".}

\item{penalty}{One of "gLASSO", "gSCAD" and "gMCP".}

\item{stepsize}{Stepsize for group coordinate descent.}

\item{tol}{Convergence tolerance.}

\item{maxit}{Maximum number of iterations.}

\item{b1}{ADAM hyperparameter, default set to 0.9.}

\item{b2}{ADAM hyperparameter, default set to 0.999.}

\item{e}{ADAM hyperparameter, default set to 1e-8.}
}
\description{
The ADAM Algorithm for GLM with gLASSO, gSCAD, gMCP penalties (experimental)
}
