% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gcd_grpreg}
\alias{gcd_grpreg}
\title{The Group Coordinate Descent Algorithm for GLM with gLASSO, gSCAD, gMCP penalties}
\usage{
gcd_grpreg(X, y, beta, lambda, a, group, family, penalty, tol, maxit)
}
\arguments{
\item{X}{Orthonormalized design matrix.}

\item{y}{The response vector.}

\item{beta}{The model coefficients.}

\item{lambda}{The penalization parameter for gMIC, e.g., 2 for AIC and long(n) for BIC.}

\item{a}{The approximation parameter for gMIC.}

\item{group}{The group structure of the model. For example, assume that X has 4 columns and group=c(1,1,2,2).
It means the first 2 features form a group of variables and the last 2 features form another group of variables.}

\item{family}{The type of glm model, should be one of "gaussian", "binomial" or "poisson".}

\item{penalty}{One of "gLASSO", "gSCAD" and "gMCP".}

\item{maxit}{Maximum number of iterations.}
}
\description{
The Group Coordinate Descent Algorithm for GLM with gLASSO, gSCAD, gMCP penalties
}
