% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gd_gmic}
\alias{gd_gmic}
\title{The gradient descent function for gMIC optimization}
\usage{
gd_gmic(X, y, a, lambda, gamma, group, family, stepsize, tol, maxit)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{The response vector.}

\item{a}{The approximation parameter for gMIC.}

\item{lambda}{The penalization parameter for gMIC, e.g., 2 for AIC and long(n) for BIC.}

\item{gamma}{The optimization parameter gamma.}

\item{group}{The group structure of the model. For example, assume that X has 4 columns and group=c(1,1,2,2).
It means the first 2 features form a group of variables and the last 2 features form another group of variables.}

\item{family}{The type of glm model, should be one of "gaussian", "binomial" or "poisson".}

\item{stepsize}{Stepsize for group coordinate descent.}

\item{tol}{Convergence tolerance.}

\item{maxit}{Maximum number of iterations.}
}
\description{
The gradient descent function for gMIC optimization
}
