% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gd_grpreg}
\alias{gd_grpreg}
\title{The Gradient Descent Algorithm for GLM with gLASSO, gSCAD, gMCP penalties}
\usage{
gd_grpreg(X, y, beta, lambda, a, group, family, penalty, stepsize, tol,
  maxit)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{The response vector.}

\item{beta}{The model coefficients.}

\item{lambda}{The penalization parameter for gMIC, e.g., 2 for AIC and long(n) for BIC.}

\item{a}{The approximation parameter for MCP or SCAD.}

\item{group}{The group structure of the model. For example, assume that X has 4 columns and group=c(1,1,2,2).
It means the first 2 features form a group of variables and the last 2 features form another group of variables.}

\item{family}{The type of glm model, should be one of "gaussian", "binomial" or "poisson".}

\item{penalty}{One of "gLASSO", "gSCAD" and "gMCP".}

\item{stepsize}{Stepsize for group coordinate descent.}

\item{tol}{Convergence tolerance.}

\item{maxit}{Maximum number of iterations.}
}
\description{
The Gradient Descent Algorithm for GLM with gLASSO, gSCAD, gMCP penalties
}
