% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_gmic}
\alias{grad_gmic}
\title{The Gradient Function for gMIC}
\usage{
grad_gmic(X, y, a, lambda, gamma, group, family)
}
\arguments{
\item{X}{The design matrix.}

\item{y}{The response vector}

\item{a}{The approximation parameter for gMIC.}

\item{lambda}{The gMIC penalization parameter.}

\item{gamma}{The optimization parameter.}

\item{group}{The group structure of the model. For example, assume that X has 4 columns and group=c(1,1,2,2).
It means the first 2 features form a group of variables and the last 2 features form another group of variables.}

\item{family}{The type of glm model, should be one of "gaussian", "binomial" or "poisson".}
}
\description{
The Gradient Function for gMIC
}
