% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grp_mic}
\alias{grp_mic}
\title{The Group MIC Penalty Function}
\usage{
grp_mic(beta, group, a)
}
\arguments{
\item{beta}{A p-dimensional vector containing the regression ceofficients.}

\item{group}{The group structure of the model. For example, assume that X has 4 columns and group=c(1,1,2,2).
It means the first 2 features form a group of variables and the last 2 features form another group of variables.}

\item{a}{The approximation parameter of the gMIC.}
}
\value{
The gMIC penalty evaluated at beta.
}
\description{
The Group MIC Penalty Function
}
