% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglik_grpreg}
\alias{loglik_grpreg}
\title{The -2*Log Likelihood Function for GLM}
\usage{
loglik_grpreg(X, y, beta, family)
}
\arguments{
\item{X}{The design matrix.}

\item{y}{The response vector.}

\item{beta}{The model coefficients.}

\item{family}{The type of glm model, should be one of "gaussian", "binomial" or "poisson".}
}
\description{
The -2*Log Likelihood Function for GLM
}
