% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoglikPenGLM.R
\name{LoglikPenGLM}
\alias{LoglikPenGLM}
\title{Compute the penalized log likelihood for GLM with MIC penalty via R function \code{glm}}
\usage{
LoglikPenGLM(gamma, group, X, y, lambda, a, family = gaussian(link =
  "identity"))
}
\arguments{
\item{group}{The group structure of the model. For example, assume that X has 4 columns and group=c(1,1,2,2).
It means the first 2 features form a group of variables and the last 2 features form another group of variables.}

\item{X}{An \eqn{n} by \eqn{p} design matrix.}

\item{y}{The \eqn{n} by 1 response vector}

\item{lambda}{The penalty parameter euqals either 2 in AIC or ln(n) in BIC (by default).
It can be specified as any value of the user's own choice.}

\item{a}{The scale parameter in the hyperbolic tangent function of the MIC penalty. By default, \eqn{a = 50}.}

\item{family}{a description of the error distribution and link function to be used in the model. It needs to be the result
of a call to a family function since \code{glm.fit} is used here. In other words, it can NOT be specified as a character
string naming a family function nor a family function. See \code{\link[stats]{glm}} and \code{\link[stats]{family}} for
details of family functions.}

\item{beta}{A p-dimensional vector containing the regression ceofficients.}
}
\value{
The value of the penalized log likelihood function evaluated at beta.
}
\description{
Compute the penalized log likelihood for GLM with MIC penalty via R function \code{glm}
}
\seealso{
\code{\link[stats]{glm}}, \code{\link[stats]{family}}
}
