% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalize.R
\name{orthogonalize}
\alias{orthogonalize}
\title{Orthogonalize the Design Matrix with Specified Group Structure}
\usage{
orthogonalize(X, group)
}
\arguments{
\item{X}{The design matrix to be orthogonalized.}

\item{group}{The group structure based on which X is orthogonalized. For example, assume that X has 4 columns and group=c(1,1,2,2).
It means the first 2 features form a group of variables and the last 2 features form another group of variables.}
}
\value{
The orthogonalized design. Assume that X_g denotes columns corresponding to the g-th group, then t(X_g)%*%(X_g) = I.
}
\description{
Orthogonalize the Design Matrix with Specified Group Structure
}
\examples{
X <- matrix(runif(1200),100,12)
group <- rep(1:3,each=4)
X <- orthogonalize(X,group)

}
